Update to new version of clang, add ptrdiff_t and etc.
Updated docker script to use the clang and available in the
distros own packet repository. Right now the python clang package
uses the libclang-17.so and this matches the versions of clang and
llvm installed with apt.

Updated ctypesgen script to accommodate the changes of
the new version of clang.

Updated the tests according to the new version of clang.